📝 Epic 정리 및 구조 잡기
